Add support for aarch64, powerpc64le, s390x and riscv64
dettrace was x86_64-only. This ports the architecture-dependent parts
to the other Linux server architectures openSUSE builds for.

Register access:
- REG_* accessors abstract struct user_regs_struct: rdi/rsi/... on
  x86_64, regs[0..5]/x8/pc on aarch64, gpr[3..8]/gpr[0]/nip on
  powerpc64le, gprs[2..7]/psw.addr on s390x and a0..a5/a7/pc on
  riscv64. Registers are read/written with PTRACE_GETREGSET/SETREGSET
  where PTRACE_GETREGS is unavailable.
- aarch64 and s390x keep the executing syscall number out of the
  general register file; it is read and written through the
  NT_ARM_SYSTEM_CALL / NT_S390_SYSTEM_CALL regsets.
- powerpc64le reports syscall errors through the cr0 summary-overflow
  bit with a positive errno; regsReturnValue/regsSetReturnValue encode
  that so the rest of the code keeps using the usual negative-errno
  convention.

Post-hook arguments:
- On every architecture except x86_64 the first argument register is
  the same register as the return value (x0/a0/r3; s390x preserves it
  in orig_gpr2), so by the time a post-hook runs arg1 has been
  overwritten. Capture all arguments at the seccomp entry stop into the
  per-tracee state and serve arg1()..arg6() from that snapshot, kept
  per tracee so it survives scheduler switches. clone/fork/vfork are
  not seccomp-intercepted, so their flags are read from the live
  registers at the PTRACE_EVENT_CLONE stop, which fires before the
  return value is written.

Missing legacy syscalls:
- syscallCompat.hpp defines the syscalls absent on the newer
  architectures (open, stat, fork, pause, time, renameat, accept, ...)
  to unique negative sentinels so the shared dispatch and handler
  classes compile. seccomp::intercept/noIntercept skip them. dettrace
  can still hijack a call via changeSystemCall(SYS_pause/SYS_time): the
  kernel treats the sentinel as sys_ni_syscall (nothing runs, like the
  intended pause/noop) while the post-hook dispatches on the sentinel,
  preserving the x86_64 semantics.

Syscall injection and vDSO:
- The post-execve injection stub is now a per-architecture syscallStub
  byte array (int3/syscall, brk/svc, ebreak/ecall, breakpoint/svc)
  with offset constants, so handleExecEvent, traceePreinitMmap and
  disableVdso are architecture-independent.
- vDSO determinization stubs for each architecture's exported symbols
  (__kernel_* on aarch64/ppc64le/s390x, __vdso_* on riscv64), plus
  riscv_hwprobe faked with -ENOSYS as both a syscall and a vDSO entry
  since it would otherwise expose nondeterministic hardware caps. All
  machine-code encodings were verified with the assembler.
- rdtsc/rdtscp/cpuid trapping (PR_SET_TSC, the SIGSEGV emulation,
  arch_prctl(ARCH_SET_CPUID)) and the legacy x86 vsyscall page handling
  are compiled out on the other architectures, which have no equivalent
  userspace instructions to determinize.

The termios2 ioctls are guarded for powerpc, which lacks them.

+++ include/seccomp.hpp
@@ -54,14 +54,14 @@ private:
    * Add system call to whitelist but no call to ptrace.
    * @param systemCall system call to add to whitelist.
    */
-  void noIntercept(uint16_t systemCall);
+  void noIntercept(int systemCall);
 
   /**
    * Add system call to whitelist but everytime it is called call ptrace to
    * intercept.
    * @param systemCall
    */
-  void intercept(uint16_t systemCall);
+  void intercept(int systemCall);
 
   /**
    * Add system call to whitelist.
@@ -69,7 +69,7 @@ private:
    * no. (useful debugging).
    * @param systemCall
    */
-  void intercept(uint16_t systemCall, bool cond);
+  void intercept(int systemCall, bool cond);
 
 public:
   /**
